requirements.txt
pyserial
windows-curses
PyQt5